OrcLib: BITSAgent: fix http error handling (changes upload behavior)
Only some of the 4xx, 5xx error were handled. This will change the upload
behavior on 'Only' mode (vs 'Overwrite') as it will not queue BITS jobs if
anything goes wrong while probing for remote file.

This is a fix to get the intended behavior: prefer no upload to multiple
for corner cases.