OrcLib: Ntfs: fix decompression for some ending block
There was multiple possible issues. The fix was done on a file which last
uncompressed block from a compression unit was corrupted. The block was
not compressed but the end was truncated.

- Last compression unit needs 65k input buffer to be sure it will be
processed. Do not use file length and processed bytes count to size
the read buffer. The compression unit is a sequence of blocks with their
own header. Even blocks which are not compressed will have a header. For
some files this could lead to truncated read buffer if it was based on
uncompressed file length.

- Before the fix, the use of 'uncomp_processed' as an offset on
'info.comp_buf', 'buffer' and 'info.uncomp_buf' seems odd.
Now each block will be processed alone and an offset is no more required
on 'info.comp_buf' and 'buffer'.